User commit

+++ src/main/java/com/entity/Student.java
@@ -1,16 +1,14 @@
 package com.example.demo.entity;
 
-
-
 public class Student {
-   private int id;
+   private Long id;
    private String name;
    private String email;
    
-   public int getId() {
+   public Long getId() {
     return id;
    }
-   public void setId(int id) {
+   public void setId(Long id) {
     this.id = id;
    }
    public String getName() {
@@ -27,7 +25,7 @@ public class Student {
    }
    
   
-   public Student(int id, String name, String email) {
+   public Student(Long id, String name, String email) {
     this.id = id;
     this.name = name;
     this.email = email;
